move src/tools directory to src/telescope_baseline/tools. Import file names are fixed.
