+++ src/telescope_baseline/tools/pipeline/detectorimage.py
@@ -1,4 +1,4 @@
-from src.tools.pipeline.simcomponent import SimComponent
+from telescope_baseline.tools.pipeline.simcomponent import SimComponent
 
 
 class DetectorImage(SimComponent):
