+++ tests/pipeline/test_upward.py
@@ -1,5 +1,5 @@
 import pytest
-from src.tools.pipeline.visitor import SimVisitor
+from telescope_baseline.tools.pipeline.visitor import SimVisitor
 
 
 class SimpleUpwardVisitor(SimVisitor):
@@ -34,8 +34,8 @@ class SimpleUpwardVisitor(SimVisitor):
 
 @pytest.fixture(scope="module")
 def fixture1():
-    from src.tools.pipeline.stellarimage import StellarImage
-    from src.tools.pipeline.detectorimage import DetectorImage
+    from telescope_baseline.tools.pipeline.stellarimage import StellarImage
+    from telescope_baseline.tools.pipeline.detectorimage import DetectorImage
     print("Pre Processing\n")
     v = SimpleUpwardVisitor()
     a = DetectorImage()
